+++ src/parallel/DomainDecomposition.h
@@ -69,7 +69,7 @@ class DomainDecomposition: public DomainDecompBase{
   double guaranteedDistance(double x, double y, double z, Domain* domain);
   
   // documentation see father class (DomainDecompBase.h)
-  int countMolecules(ParticleContainer* moleculeContainer, vector<int> &compCount);
+  unsigned long countMolecules(ParticleContainer* moleculeContainer, vector<unsigned long> &compCount);
 
   // documentation see father class (DomainDecompBase.h)
   double getBoundingBoxMin(int dimension, Domain* domain);
